requirements.txt
appnope==0.1.0
aws-psycopg2==1.1.1
awsebcli==3.15.3
backcall==0.1.0
beautifulsoup4==4.8.0
blessed==1.15.0
botocore==1.12.238
cached-property==1.5.1
cement==2.8.2
certifi==2019.9.11
chardet==3.0.4
colorama==0.3.9
coverage==4.5.4
decorator==4.4.0
Django==2.2.5
django-debug-toolbar==2.0
django-htmlmin==0.11.0
django-js-asset==1.2.2
django-mptt==0.10.0
django-ratelimit==2.0.0
docker==3.7.3
docker-compose==1.23.2
docker-pycreds==0.4.0
dockerpty==0.4.1
docopt==0.6.2
docutils==0.15.2
future==0.16.0
html5lib==1.0.1
idna==2.7
ipython==7.8.0
ipython-genutils==0.2.0
jedi==0.15.1
jmespath==0.9.4
jsonschema==2.6.0
parso==0.5.1
pathspec==0.5.9
pexpect==4.7.0
pickleshare==0.7.5
prompt-toolkit==2.0.9
ptyprocess==0.6.0
Pygments==2.4.2
python-dateutil==2.8.0
pytz==2019.2
PyYAML==3.13
requests==2.20.1
semantic-version==2.5.0
six==1.11.0
soupsieve==1.9.3
sqlparse==0.3.0
termcolor==1.1.0
texttable==0.9.1
traitlets==4.3.2
urllib3==1.24.3
wcwidth==0.1.7
webencodings==0.5.1
websocket-client==0.56.0
